Backend/ml_research/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.21.0
pillow>=9.0.0
opencv-python>=4.6.0
pytesseract>=0.3.10
numpy>=1.22.0
tqdm>=4.64.0
